[bug fix] Update file path in code chunking process
Update the file path used in the code_chunker.chunk_file() method to use the full path by joining the directory with the relative file path. This ensures that the correct file is accessed when chunking code files.

Notes: The change addresses a potential issue where the code chunker might not be able to locate files correctly if the current working directory is different from where the files are stored. By using os.path.join(directory, doc[\"file_path\"]), we create an absolute path that should work regardless of the current working directory.